use "Partition unique GUID" from gdisk / GPT
Ok, so the only way I was able to get this was using gdisk, and then only
after I switched from fdisk to parted for creating the partition.  (The
latter _seems_ way better, 'cause it takes scripted commands as arguments
and does atomic operations, but there may be weirdness I don't understand.)

I don't really understand the different between this value and the one you can
get with blkid and tune2fs (which seem to give me the same thing).  I'm kind
of awed there's no good way to just get this from a reasonable CLI interface.

The state of tooling and documentation around this stuff is _completely awful_.